primeros pasos de autenticacion realizados.

+++ prisma/migrations/20240518012851_init/migration.sql
@@ -4,7 +4,7 @@ CREATE TABLE "Product" (
     "name" TEXT NOT NULL,
     "price" REAL NOT NULL,
     "sold" INTEGER NOT NULL,
-    "discount" INTEGER,
+    "discount" REAL,
     "available" TEXT NOT NULL DEFAULT 'disponible',
     "description" TEXT,
     "createdAt" DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
